Move Table story next to Table component
